fix: namespaced matched route
